% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brclimr-package.R
\docType{package}
\name{brclimr-package}
\alias{brclimr}
\alias{brclimr-package}
\title{brclimr: Fetch zonal weather zonal indicators for Brazilian municipalities}
\description{
Fetch zonal weather zonal indicators for Brazilian municipalities from brclim project.
}
\author{
\strong{Maintainer}: Raphael Saldanha \email{raphael.de-freitas-saldanha@inria.fr} (\href{https://orcid.org/0000-0003-0652-8466}{ORCID})

}
\keyword{internal}
