% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brdwgd_data.R
\docType{data}
\name{brdwgd_data}
\alias{brdwgd_data}
\title{BR-DWGD zonal statistics data}
\format{
An object of class \code{list} of length 7.
}
\usage{
brdwgd_data
}
\description{
A list with BR-DWGD zonal indicators data.
}
\keyword{datasets}
