% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch data from given parameters}
\usage{
fetch_data(code_muni, product, indicator, statistics, date_start, date_end)
}
\arguments{
\item{code_muni}{IBGE municipality code number with 7 digits.}

\item{product}{Zonal indicator product. Options are: \code{brdwgd}, \code{terraclimate}.}

\item{indicator}{Indicator name from the product. Check the available indicators with the \link{product_info} function.}

\item{statistics}{Statistics name from the indicator. Check the available statistics with the \link{product_info} function.}

\item{date_start}{A date.}

\item{date_end}{A date, >= than \code{date_start}.}
}
\value{
A \code{data.frame} with date and value columns.
}
\description{
Fetch data from given parameters
}
\examples{
\donttest{
fetch_data(
    code_muni = 3304557,
    product = "brdwgd",
    indicator = "tmax",
    statistics = "mean",
    date_start = as.Date("2008-01-01"),
    date_end = as.Date("2008-01-10")
 )
 }
}
