% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_info.R
\name{product_info}
\alias{product_info}
\title{Print product information, indicators acronyms and available statistics}
\usage{
product_info(product = "brdwgd")
}
\arguments{
\item{product}{Zonal indicator product. Options are: \code{brdwgd}, \code{terraclimate}.}
}
\value{
Printed list tree on the console.
}
\description{
Print product information, indicators acronyms and available statistics
}
\examples{
product_info()
}
