% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terraclimate_data.R
\docType{data}
\name{terraclimate_data}
\alias{terraclimate_data}
\title{TerraClimate zonal indicators data}
\format{
An object of class \code{list} of length 14.
}
\usage{
terraclimate_data
}
\description{
A list with TerraClimate zonal indicators data.
}
\keyword{datasets}
