% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_info.R
\name{product_info}
\alias{product_info}
\title{Print product indicators and statistics}
\usage{
product_info(product = "brdwgd")
}
\arguments{
\item{product}{Zonal indicator product. Currently, only \code{brdwgd}.}
}
\value{
Printed messages on the console.
}
\description{
The information is printed in the following order: indicator name, indicator acronyms in brackets, unit in parenthesis, available statistics, date range in brackets.
}
\examples{
product_info()
}
